Split Code::Pos()/Code::Style() functions into a file
